LPS-167154 Create a new field tabs directory structure and move all components related to the new Basic Info tab there
